+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2rc1] - 2026-07-19
+
+Release candidate for TestPyPI (same content as 0.1.2 final).
+
 ## [0.1.2] - 2026-07-19
 
 ### Fixed